[CSS Shaders] Make CSS Shaders render to texture framebuffers https://bugs.webkit.org/show_bug.cgi?id=85113
Reviewed by Dean Jackson.

Source/WebCore:

Added the required members to store the framebuffer, the texture and the depth buffer. Also removed the m_drawingBuffer
because context->readPixels doesn't require one anymore.

ReadPixels is not flipping the result, so I've also removed the flipping projection matrix that was specific only to Chromium.

Added a test to check that colors are not switched and the resulting image is not mirrored.

Test: css3/filters/custom/effect-color-check.html

* platform/graphics/filters/FECustomFilter.cpp:
(WebCore::FECustomFilter::FECustomFilter):
(WebCore::FECustomFilter::~FECustomFilter):
(WebCore):
(WebCore::FECustomFilter::deleteRenderBuffers):
(WebCore::FECustomFilter::platformApplySoftware):
(WebCore::FECustomFilter::initializeContext):
(WebCore::FECustomFilter::resizeContext):
(WebCore::FECustomFilter::bindProgramAndBuffers):
* platform/graphics/filters/FECustomFilter.h:
(FECustomFilter):

LayoutTests:

Added a test to check that colors are not switched and the resulting image is not mirrored.

* css3/filters/custom/effect-color-check-expected.png: Added.
* css3/filters/custom/effect-color-check-expected.txt: Added.
* css3/filters/custom/effect-color-check.html: Added.
* css3/filters/resources/color-fill.fs: Added.
* platform/chromium-linux/css3/filters/custom/custom-filter-shader-cache-expected.png:
* platform/chromium-linux/css3/filters/custom/effect-custom-combined-missing-expected.png:
* platform/chromium-linux/css3/filters/custom/effect-custom-expected.png:
* platform/chromium-mac/css3/filters/custom/effect-color-check-expected.png: Added.
* platform/chromium/test_expectations.txt:


Identifier: 103829@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@116795 268f45cc-cd09-0410-ab3c-d52691b4dbfc