[BlackBerry] Allow the platform media player to determine the media element's paused/playing status https://bugs.webkit.org/show_bug.cgi?id=86235
Patch by Max Feil <mfeil@rim.com> on 2012-05-12
Reviewed by George Staikos.

Source/WebCore:

The platform media player needs to know when the HTMLMediaElement
is not paused. This is to address problems when switching
source element, which causes the destruction of the old
MediaPlayerPrivate object and construction of a new one. The
new one must resume playing ASAP if the old one was playing.

Test: media/media-continues-playing-after-replace-source.html

* platform/graphics/blackberry/MediaPlayerPrivateBlackBerry.cpp:
(WebCore::MediaPlayerPrivate::isElementPaused):
(WebCore):
* platform/graphics/blackberry/MediaPlayerPrivateBlackBerry.h:
(MediaPlayerPrivate):

LayoutTests:

This test replaces the source element of a playing media object
and makes sure the new source element begins playing.

* media/media-continues-playing-after-replace-source-expected.txt: Added.
* media/media-continues-playing-after-replace-source.html: Added.

Identifier: 103890@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@116858 268f45cc-cd09-0410-ab3c-d52691b4dbfc