[chromium] Prevent KeyCodeConversionAndroid from breaking on next NDK roll https://bugs.webkit.org/show_bug.cgi?id=86415
Patch by Alexandre Elias <aelias@google.com> on 2012-05-14
Reviewed by Adam Barth.

The new enum values added in this file have been added in recent
versions of the NDK, so this file will fail to compile with duplicate
enum error when we roll NDK to r7c or higher.  A workaround is
to move our enum values to the WebCore namespace.  When the NDK
upgrade is complete, we can delete them.

No new tests. (No-op change.)

* platform/chromium/KeyCodeConversionAndroid.cpp:
(WebCore):


Identifier: 104013@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@117022 268f45cc-cd09-0410-ab3c-d52691b4dbfc