Make StyleResolver::applyMatchedProperties and ::applyProperties use enum template parameter instead of bool. https://bugs.webkit.org/show_bug.cgi?id=86424
Reviewed by Dimitri Glazkov.

This patch changes from using a boolean template parameter for StyleResolver::applyMatchedProperties and
StyleResolver::applyProperties functions. The motivation for this change is that it paves the way for
CSS Variables to make the value tri-state, which is a requirement because variable definitions must occur
before high priority CSS properties such that they can be referenced by the latter. This change affects
only the type signatures of those functions, and not their behavior.

* css/StyleResolver.cpp:
(WebCore::StyleResolver::collectMatchingRulesForList):
* css/StyleResolver.h:


Identifier: 104025@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@117034 268f45cc-cd09-0410-ab3c-d52691b4dbfc