WKContextSetCacheModel(contextref, kWKCacheModelDocumentViewer) doesn't prevent pages from going into the PageCache. https://bugs.webkit.org/show_bug.cgi?id=85891
Reviewed by Brady Eidson.

Source/WebCore:

Do not claim that it is possible to cache a page when the page cache has a capacity of 0.

Covered by the modified DOMWindowExtensionNoCache WK2 API test.

* history/PageCache.cpp:
(WebCore::PageCache::canCache):
Check m_capacity.
* history/PageCache.h:

* loader/FrameLoader.cpp:
(WebCore::FrameLoader::commitProvisionalLoad):
canCache is no longer static on PageCache.

Tools:

Set the cache model in DOMWindowExtensionNoCache so that the page cache is not used. This is
better than using pages with unload handlers because we may find a way to make pages with
unload handlers cacheable in the future.

* TestWebKitAPI/TestWebKitAPI.xcodeproj/project.pbxproj:
Remove simple-iframe-unload.html and simple-unload.html, since they are no longer used.

* TestWebKitAPI/Tests/WebKit2/DOMWindowExtensionNoCache.cpp:
(TestWebKitAPI::TEST):
Set the cache model to kWKCacheModelDocumentViewer.

* TestWebKitAPI/Tests/WebKit2/simple-iframe-unload.html: Removed.
* TestWebKitAPI/Tests/WebKit2/simple-unload.html: Removed.

Identifier: 104089@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@117112 268f45cc-cd09-0410-ab3c-d52691b4dbfc