Skia does not respect a specified InterpolationQuality https://bugs.webkit.org/show_bug.cgi?id=86249
Patch by Keyar Hood <keyar@chromium.org> on 2012-05-15
Reviewed by Stephen White.

Source/WebCore:

The added functionality is not exposed to higher layers of webkit.
Tests will be added that will exercise these changes when bug 82804 is
fixed.

* platform/graphics/GraphicsContext.cpp:
(WebCore::GraphicsContext::drawImage):
(WebCore::GraphicsContext::drawImageBuffer):
Code to have the useLowQualityScale cause the InterpolationQuality be
be set to low for Chromium but remain as none for other platforms.
* platform/graphics/skia/ImageSkia.cpp:
(WebCore::limitResamplingMode): Added
(WebCore):
(WebCore::paintSkBitmap):
(WebCore::Image::drawPattern):
We now limit the resampling choice based on what InterpolationQuality
is set as. InterpolationNone restricts resampling to RESAMPLE_NONE,
InterpolationLow and InterpolationMedium restricts resampling to
RESAMPLE_LINEAR. InterpolationHigh and InterpolationDefault do not
change the resampling.

Furthermore, the choice on how to set the filter bitmap flag in
paintSkBitmap was made to be consistent with that in
Image::drawPattern.

LayoutTests:

Now expecting the following to tests to have an IMAGE failure:
svg/custom/pointer-events-image-css-transform.svg
svg/custom/pointer-events-image.svg
* platform/chromium/test_expectations.txt:

Identifier: 104124@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@117187 268f45cc-cd09-0410-ab3c-d52691b4dbfc