remove bit rotted comment about css properties https://bugs.webkit.org/show_bug.cgi?id=86750
Reviewed by Alexis Menard.

It's not clear to me what the first block was even referring to.
Aural doesn't seem to exist anymore. It might be worth keeping,
"All the CSS properties are not supported by the renderer at the
moment.", but it seems obvious to me.

No new tests, just removing a comment.

* css/CSSParser.cpp:
(WebCore::CSSParser::parseValue):


Identifier: 104375@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@117468 268f45cc-cd09-0410-ab3c-d52691b4dbfc