[EFL] Do not enable Web Audio support by default at build time. https://bugs.webkit.org/show_bug.cgi?id=86982
Reviewed by Philippe Normand.

Follow-up to r115800. Even though Web Audio support is disabled by
default in build-webkit, enabling it in OptionsEfl.cmake will turn
it on when building the port without build-webkit.

Since Web Audio support for the EFL port does not seem to be fully
working yet, it is better to disable it in all cases.

* Source/cmake/OptionsEfl.cmake: Remove option override for
ENABLE_WEB_AUDIO.


Identifier: 104696@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@117807 268f45cc-cd09-0410-ab3c-d52691b4dbfc