[Qt] canvas/philip/tests/2d.path.stroke.overlap.html fails with Qt-4.6.2 or higher https://bugs.webkit.org/show_bug.cgi?id=40362
Reviewed by Noam Rosenthal.

Source/WebCore:

Before drawing the stroke from a certain path, it should make the union of the stroke areas.
The previous implementation didn't consider that so the intersection was painted twice.
By using QPainterPathStroker, we create a QPainterPath on top of the platform path, and
then we fill it with the expected brush. This way we avoid repainting areas.

* platform/graphics/qt/GraphicsContextQt.cpp:
(WebCore::fillPathStroke):
(WebCore):
(WebCore::GraphicsContext::strokePath):

LayoutTests:

* platform/qt/Skipped: Unskip test after fix.


Identifier: 104866@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@118020 268f45cc-cd09-0410-ab3c-d52691b4dbfc