WebKit spends ~20% of time in HTMLTextAreaElement::defaultValue() when opening a review page https://bugs.webkit.org/show_bug.cgi?id=86813
Reviewed by Eric Seidel.

Replace String and calls to String::operator+= by StringBuilder and calls to append respectively.
While there are other optimizations we can implement such as not replacing all nodes in the shadow DOM,
or not serializing the value when detecting whether the value had changed, they require considerable
amount of refactoring and this change alone seems to buy us a lot so let us defer those 'til later time.

Also avoid the redundant call to setValue in childrenChanged when the value is dirty.
This should halve the time we spend in setting the value.

* html/HTMLTextAreaElement.cpp:
(WebCore::HTMLTextAreaElement::childrenChanged):
(WebCore::HTMLTextAreaElement::defaultValue):


Identifier: 104961@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@118149 268f45cc-cd09-0410-ab3c-d52691b4dbfc