Web Inspector: Debugger should correctly show catch scope https://bugs.webkit.org/show_bug.cgi?id=87197
Reviewed by Vsevolod Vlasov.

Source/WebCore:

* bindings/v8/DebuggerScript.js: In case of Catch scope the scope object
is a synthetic one which means that it contains all standard object properties.
When passing this object further we need to leave only properties available in
the catch block(only variable thar stores the exception in fact).

LayoutTests:

Added test for "catch" block to the debugger test.

* inspector/debugger/debugger-expand-scope-expected.txt:
* inspector/debugger/debugger-expand-scope.html:
* platform/chromium/inspector/debugger/debugger-expand-scope-expected.txt:

Identifier: 104972@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@118161 268f45cc-cd09-0410-ab3c-d52691b4dbfc